+++ tests/test_atomsmithy.py
@@ -129,5 +129,5 @@ def test_push_cutoff():
 def test_random_box():
     rng = np.random.default_rng(42)
     atpos = build_random_cell(10.0, 32, 3.0, rng=rng)
-    nlist = build_nlist_simple(atpos, [10.0, 10.0, 10.0], rcut=4.0)
+    nlist = build_nlist_simple(atpos, [10.0, 10.0, 10.0], rmax=4.0)
     assert nlist["dist"].min() > 2.0