+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- Fix bug in `PairwiseForceField` class: use `rmax` only, never `rcut`.
+- Fix version import.
+
+
 ## [0.2.0] - 2024-10-07
 
 ## Changed